Multiple branch test
My goal is to have everything essentially read-only style.  I want
to use this for a website, so the intention is that if you have
modified a branch, you need to make a new class that points to it,
and you'll be able to resolve the modified data.

This will also make the app load much faster because it won't call
git commands multiple times for one go.  In order to remove caching,
you'll have to create a new object.