dzil build with versioning
This sets up dzil build.  In order to use the dzil build process,
just do a dzil build and then we have something that is ready for
release.  I decided not to use autoversioning for now because I
really don't understand how it kept on ending up with a 1.4...
version number, and will figure out autoversioning later.